Move dimension label headers up a level
