+++ test/CMakeLists.txt
@@ -258,7 +258,7 @@ endif()
 if (TILEDB_EXPERIMENTAL_FEATURES)
   list(APPEND TILEDB_UNIT_TEST_SOURCES
     src/test-capi-array-many-dimension-labels.cc
-    src/test-capi-dimension_labels.cc
+    src/test-capi-dimension-label.cc
     src/test-capi-dimension-label-encrypted.cc
     src/test-capi-dense-array-dimension-label.cc
     src/test-capi-dense-array-dimension-label-var.cc
