+++ tiledb/CMakeLists.txt
@@ -78,7 +78,7 @@ endif()
 
 if (TILEDB_EXPERIMENTAL_FEATURES)
   list(APPEND TILEDB_PUBLIC_HEADERS
-    ${TILEDB_CORE_INCLUDE_DIR}/tiledb/sm/c_api/experimental/tiledb_dimension_label.h
+    ${TILEDB_CORE_INCLUDE_DIR}/tiledb/sm/c_api/tiledb_dimension_label.h
   )
 endif()
 
@@ -120,6 +120,11 @@ if (TILEDB_CPP_API)
     ${TILEDB_CORE_INCLUDE_DIR}/tiledb/sm/cpp_api/arrowio
     ${TILEDB_CORE_INCLUDE_DIR}/tiledb/sm/cpp_api/arrow_io_impl.h
   )
+  if (TILEDB_EXPERIMENTAL_FEATURES)
+    list(APPEND TILEDB_PUBLIC_HEADERS
+      ${TILEDB_CORE_INCLUDE_DIR}/tiledb/sm/cpp_api/dimension_label.h
+    )
+  endif()
 else()
   message(STATUS "TileDB C++ API is not built.")
 endif()
@@ -289,7 +294,7 @@ list(APPEND TILEDB_CORE_SOURCES ${TILEDB_API_SOURCES})
 
 if (TILEDB_EXPERIMENTAL_FEATURES)
   list(APPEND TILEDB_CORE_SOURCES
-    ${TILEDB_CORE_INCLUDE_DIR}/tiledb/sm/c_api/experimental/tiledb_dimension_label.cc)
+    ${TILEDB_CORE_INCLUDE_DIR}/tiledb/sm/c_api/tiledb_dimension_label.cc)
 endif()
